chore: Yet more cleanup of dead code / cruft
