+++ src/features/transfer/utils.ts
@@ -1,3 +1,69 @@
+import ConfirmedIcon from '../../images/icons/confirmed-icon.svg';
+import DeliveredIcon from '../../images/icons/delivered-icon.svg';
+import ErrorCircleIcon from '../../images/icons/error-circle.svg';
+import { FinalTransferStatuses, SentTransferStatuses, TransferStatus } from './types';
+
+export function getTransferStatusLabel(
+  status: TransferStatus,
+  connectorName: string,
+  isPermissionlessRoute: boolean,
+  isAccountReady: boolean,
+) {
+  let statusDescription = '...';
+  if (!isAccountReady && !FinalTransferStatuses.includes(status))
+    statusDescription = 'Please connect wallet to continue';
+  else if (status === TransferStatus.Preparing)
+    statusDescription = 'Preparing for token transfer...';
+  else if (status === TransferStatus.CreatingTxs) statusDescription = 'Creating transactions...';
+  else if (status === TransferStatus.SigningApprove)
+    statusDescription = `Sign approve transaction in ${connectorName} to continue.`;
+  else if (status === TransferStatus.ConfirmingApprove)
+    statusDescription = 'Confirming approve transaction...';
+  else if (status === TransferStatus.SigningTransfer)
+    statusDescription = `Sign transfer transaction in ${connectorName} to continue.`;
+  else if (status === TransferStatus.ConfirmingTransfer)
+    statusDescription = 'Confirming transfer transaction...';
+  else if (status === TransferStatus.ConfirmedTransfer)
+    if (!isPermissionlessRoute)
+      statusDescription = 'Transfer transaction confirmed, delivering message...';
+    else
+      statusDescription =
+        'Transfer confirmed, the funds will arrive when the message is delivered.';
+  else if (status === TransferStatus.Delivered)
+    statusDescription = 'Delivery complete, transfer successful!';
+  else if (status === TransferStatus.Failed)
+    statusDescription = 'Transfer failed, please try again.';
+
+  return statusDescription;
+}
+
+export function isTransferSent(status: TransferStatus) {
+  return SentTransferStatuses.includes(status);
+}
+
+export function isTransferFailed(status: TransferStatus) {
+  return status === TransferStatus.Failed;
+}
+
+export const STATUSES_WITH_ICON = [
+  TransferStatus.Delivered,
+  TransferStatus.ConfirmedTransfer,
+  TransferStatus.Failed,
+];
+
+export function getIconByTransferStatus(status: TransferStatus) {
+  switch (status) {
+    case TransferStatus.Delivered:
+      return DeliveredIcon;
+    case TransferStatus.ConfirmedTransfer:
+      return ConfirmedIcon;
+    case TransferStatus.Failed:
+      return ErrorCircleIcon;
+    default:
+      return ErrorCircleIcon;
+  }
+}
+
 import {
   ChainMap,
   CoreAddresses,
